Added support for >10 options
