+++ package.json
@@ -8,7 +8,6 @@
     "prestart": "npm run migrate",
     "dev": "nodemon index.js",
     "tunnel": "ngrok http 3000",
-    "test": "jest",
     "migrate": "if [ $DATABASE_URL ]; then sequelize db:migrate --url $DATABASE_URL ; else echo; fi"
   },
   "engines": {
@@ -17,19 +16,15 @@
   "dependencies": {
     "dotenv": "^6.1.0",
     "express": "^4.13.4",
-    "fetch-mock": "^7.2.5",
-    "node-fetch": "^2.3.0",
+    "morgan": "^1.9.1",
     "pg": "^7.7.1",
     "pg-hstore": "^2.3.2",
     "sequelize": "^4.41.2",
     "sequelize-cli": "^5.4.0",
-    "sqlite3": "^4.0.4"
+    "sqlite3": "^4.0.4",
+    "axios": "^0.18.0"
   },
   "devDependencies": {
-    "axios": "^0.18.0",
-    "jest": "^23.6.0",
-    "moxios": "^0.4.0",
-    "nodemon": "^1.18.7",
-    "supertest": "^3.3.0"
+    "nodemon": "^1.18.7"
   }
 }
