+++ utils.js
@@ -1,5 +1,11 @@
+function chunkArray (array, limit, res = []) {
+  if (!limit) throw new Error('No chunk size defined')
+  if (array.length === 0) return res
+  return chunkArray(array.slice(limit, array.length), limit, [...res, array.slice(0, limit)])
+}
+
 function escapeRegExp (str) {
-  return str.replace(/([.*+?^=!:${}()|\[\]\/\\])/g, "\\$1")
+  return str.replace(/([.*+?^=!:${}()|\[\]\/\\])/g, '\\$1')
 }
 
 function replaceAll (str, find, replace) {
@@ -11,6 +17,7 @@ function cleanDoubleQuotes (text) {
 }
 
 module.exports = {
+  chunkArray,
   replaceAll,
   cleanDoubleQuotes
 }
